# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobRead": {
        "properties": {
          "cancel_requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Requested At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failed_count": {
            "title": "Failed Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "skipped_count": {
            "title": "Skipped Count",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "succeeded_count": {
            "title": "Succeeded Count",
            "type": "integer"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "priority",
          "payload",
          "result",
          "total_count",
          "succeeded_count",
          "failed_count",
          "skipped_count",
          "cancel_requested_at",
          "started_at",
          "finished_at",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead",
        "type": "object"
      },
      "MailAttachmentRead": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_inline": {
            "title": "Is Inline",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "content_type",
          "size",
          "is_inline"
        ],
        "title": "MailAttachmentRead",
        "type": "object"
      },
      "MailDetailRead": {
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/MailAttachmentRead"
            },
            "title": "Attachments",
            "type": "array"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "body_type": {
            "title": "Body Type",
            "type": "string"
          },
          "cc": {
            "items": {
              "type": "string"
            },
            "title": "Cc",
            "type": "array"
          },
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "id_mode": {
            "title": "Id Mode",
            "type": "string"
          },
          "is_read": {
            "title": "Is Read",
            "type": "boolean"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "received_at": {
            "title": "Received At",
            "type": "string"
          },
          "recipients": {
            "items": {
              "type": "string"
            },
            "title": "Recipients",
            "type": "array"
          },
          "sender": {
            "title": "Sender",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "id",
          "folder",
          "subject",
          "sender",
          "recipients",
          "cc",
          "received_at",
          "is_read",
          "body",
          "body_type",
          "attachments",
          "id_mode",
          "method"
        ],
        "title": "MailDetailRead",
        "type": "object"
      },
      "MailPageRead": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MailSummaryRead"
            },
            "title": "Items",
            "type": "array"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "partial_errors": {
            "additionalProperties": true,
            "title": "Partial Errors",
            "type": "object"
          }
        },
        "required": [
          "items",
          "has_more",
          "method"
        ],
        "title": "MailPageRead",
        "type": "object"
      },
      "MailSummaryRead": {
        "properties": {
          "body_preview": {
            "title": "Body Preview",
            "type": "string"
          },
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "has_attachments": {
            "title": "Has Attachments",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "id_mode": {
            "title": "Id Mode",
            "type": "string"
          },
          "is_read": {
            "title": "Is Read",
            "type": "boolean"
          },
          "received_at": {
            "title": "Received At",
            "type": "string"
          },
          "recipients": {
            "items": {
              "type": "string"
            },
            "title": "Recipients",
            "type": "array"
          },
          "sender": {
            "title": "Sender",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "id",
          "folder",
          "subject",
          "sender",
          "recipients",
          "received_at",
          "is_read",
          "has_attachments",
          "body_preview",
          "id_mode"
        ],
        "title": "MailSummaryRead",
        "type": "object"
      },
...
      "RetentionPolicyRead": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "folders": {
            "items": {
              "type": "string"
            },
            "title": "Folders",
            "type": "array"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "max_age_days": {
            "title": "Max Age Days",
            "type": "integer"
          },
          "max_messages": {
            "title": "Max Messages",
            "type": "integer"
          },
          "retain_bodies": {
            "title": "Retain Bodies",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "enabled",
          "retain_bodies",
          "folders",
          "max_messages",
          "max_age_days",
          "last_synced_at",
          "created_at",
          "updated_at"
        ],
        "title": "RetentionPolicyRead",
        "type": "object"
      },
      "RetentionPolicyWrite": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "folders": {
            "items": {
              "enum": [
                "inbox",
                "junkemail"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Folders",
            "type": "array"
          },
          "max_age_days": {
            "default": 30,
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Max Age Days",
            "type": "integer"
          },
          "max_messages": {
            "default": 1000,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Max Messages",
            "type": "integer"
          },
          "retain_bodies": {
            "default": false,
            "title": "Retain Bodies",
            "type": "boolean"
          }
        },
        "title": "RetentionPolicyWrite",
        "type": "object"
      },
      "RetentionStatsRead": {
        "properties": {
          "account_count": {
            "title": "Account Count",
            "type": "integer"
          },
          "body_count": {
            "title": "Body Count",
            "type": "integer"
          },
          "estimated_bytes": {
            "title": "Estimated Bytes",
            "type": "integer"
          },
          "message_count": {
            "title": "Message Count",
            "type": "integer"
          }
        },
        "required": [
          "account_count",
          "message_count",
          "body_count",
          "estimated_bytes"
        ],
        "title": "RetentionStatsRead",
        "type": "object"
      },
      "RetentionSyncJobCreate": {
        "properties": {
          "account_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 500,
            "title": "Account Ids",
            "type": "array"
          },
          "limit": {
            "default": 500,
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          }
        },
        "title": "RetentionSyncJobCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Smart Email Manager API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/accounts/{account_id}/mail": {
      "get": {
        "operationId": "list_mail",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "default": "inbox",
              "enum": [
                "inbox",
                "junkemail",
                "deleteditems",
                "all"
              ],
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "graph",
                "imap"
              ],
              "title": "Method",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailPageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mail",
        "tags": [
          "mail"
        ]
      }
    },
...
    "/api/v1/retention/accounts/{account_id}/mail": {
      "get": {
        "operationId": "list_retained_mail",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "default": "inbox",
              "enum": [
                "inbox",
                "junkemail",
                "all"
              ],
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailPageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cached Mail",
        "tags": [
          "mail retention"
        ]
      }
    },
    "/api/v1/retention/accounts/{account_id}/mail/{message_id}": {
      "get": {
        "operationId": "get_retained_mail_detail",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "default": "inbox",
              "enum": [
                "inbox",
                "junkemail"
              ],
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cached Detail",
        "tags": [
          "mail retention"
        ]
      }
    },
    "/api/v1/retention/accounts/{account_id}/policy": {
      "get": {
        "operationId": "get_retention_policy",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy",
        "tags": [
          "mail retention"
        ]
      },
      "put": {
        "operationId": "write_retention_policy",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionPolicyWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Policy",
        "tags": [
          "mail retention"
        ]
      }
    },
    "/api/v1/retention/cache": {
      "delete": {
        "operationId": "clear_retained_mail",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cache",
        "tags": [
          "mail retention"
        ]
      }
    },
    "/api/v1/retention/policies": {
      "get": {
        "operationId": "list_retention_policies",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RetentionPolicyRead"
                  },
                  "title": "Response List Retention Policies",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policies",
        "tags": [
          "mail retention"
        ]
      }
    },
    "/api/v1/retention/stats": {
      "get": {
        "operationId": "get_retention_stats",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionStatsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stats",
        "tags": [
          "mail retention"
        ]
      }
    },
    "/api/v1/retention/sync-jobs": {
      "post": {
        "operationId": "create_retention_sync_job",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionSyncJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Sync Job",
        "tags": [
          "mail retention"
        ]
      }
    },
...
  }
}